better implement errors

+++ src/structures/Channel.js
@@ -2,6 +2,7 @@
 
 const Snowflake = require('../util/Snowflake');
 const Base = require('./Base');
+const { Error } = require('../errors');
 const { ChannelTypes } = require('../util/Constants');
 
 /**
@@ -82,6 +83,7 @@ class Channel extends Base {
    *   .catch(console.error);
    */
   delete() {
+    if(this.type === 'group') return Promise.reject(new Error('DELETE_GROUP_DM_CHANNEL'));
     return this.client.api.channels(this.id).delete().then(() => this);
   }
 
@@ -90,6 +92,7 @@ class Channel extends Base {
    * @returns {Promise<Channel>}
    */
   fetch() {
+    if(this.type === 'group') return Promise.reject(new Error('FETCH_GROUP_DM_CHANNEL'));
     return this.client.channels.fetch(this.id, true);
   }
